refactor: Text size를 변경하다
