+++ src/components/common/header/UserCartInfo.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { Text } from '../Text';
+import { Text } from '../Text/Text';
 
 const UserCartInfo = () => {
   return (
